refactor: modularize recipe ingredients step
