--- src/components/recipe/components/RecipeForm/components/index.ts
@@ -1,6 +1,0 @@
-// src/components/recipe/components/RecipeForm/components/index.ts
-
-export { IngredientSelector } from './IngredientSelector';
-export { IngredientTable } from './IngredientTable';
-export { ConversionInfoPanel } from './ConversionInfoPanel';
-export { IngredientSummary } from './IngredientSummary';
