+++ src/components/recipe/components/RecipeForm/hooks/index.ts
@@ -1,5 +1,4 @@
 // src/components/recipe/components/RecipeForm/hooks/index.ts
 
-export { useIngredientSelection } from './useIngredientSelection';
-export { useIngredientCalculation } from './useIngredientCalculation';  
+export { useIngredientCalculation } from './useIngredientCalculation';
 export { useIngredientFormManager } from './useIngredientFormManager';
