+++ src/components/recipe/components/RecipeForm/index.tsx
@@ -17,7 +17,7 @@ import { toast } from 'sonner';
 import { logger } from '@/utils/logger';
 // Form components
 import BasicInfoStep from './BasicInfoStep';
-import IngredientsStep from './IngredientsStepRefactored';
+import IngredientsStep from './IngredientsStep';
 
 // Lazy load CostCalculationStep
 const CostCalculationStep = React.lazy(() =>
