+++ src/components/recipe/components/RecipeForm/ingredients-step/components/index.ts
@@ -1,0 +1,7 @@
+// src/components/recipe/components/RecipeForm/ingredients-step/components/index.ts
+
+export { IngredientSelector } from './IngredientSelector';
+export { IngredientTable } from './IngredientTable';
+export { ConversionInfo } from './ConversionInfo';
+export { IngredientSummary } from './IngredientSummary';
+export { IngredientFormFields } from './IngredientFormFields';
