+++ src/components/recipe/components/RecipeForm/ingredients-step/hooks/index.ts
@@ -1,0 +1,6 @@
+// src/components/recipe/components/RecipeForm/ingredients-step/hooks/index.ts
+
+export { useIngredientConversion } from './useIngredientConversion';
+export { useIngredientSelection } from './useIngredientSelection';
+export { useIngredientValidation } from './useIngredientValidation';
+export type { IngredientConversionApi } from './useIngredientConversion';
