+++ src/config/codeSplitting.ts
@@ -95,8 +95,8 @@ export const RecipeComponents = {
       .catch(() => ({ default: () => React.createElement('div', { className: 'p-4 text-center text-red-500' }, 'Gagal memuat filter resep') }))
   ),
   
-  IngredientsStep: React.lazy(() => 
-    import(/* webpackChunkName: "ingredients-step" */ '@/components/recipe/components/RecipeForm/IngredientsStepRefactored')
+  IngredientsStep: React.lazy(() =>
+    import(/* webpackChunkName: "ingredients-step" */ '@/components/recipe/components/RecipeForm/IngredientsStep')
       .catch(() => ({ default: () => React.createElement('div', { className: 'p-4 text-center text-red-500' }, 'Gagal memuat langkah bahan') }))
   )
 };
